requirements.txt
future~=0.18.2
Pillow~=9.1.1
dlib~=19.24.0
tensorflow~=2.5.3
tensorflow-gpu~=2.4.0
opencv-python~=4.6.0.66
numpy~=1.19.5
Flask~=2.2.2
gevent~=21.12.0